# Implement WAL replay and markers for `loki.write`