--- a/component/common/loki/client/batch.go
+++ b/component/common/loki/client/batch.go
@@ -19,6 +19,12 @@ const (
 	errMaxStreamsLimitExceeded = "streams limit exceeded, streams: %d exceeds limit: %d, stream: '%s'"
 )
 
+// SentDataMarkerHandler is a slice of the MarkerHandler interface, that the batch interacts with to report the event that
+// all data in the batch has been delivered or a client failed to do so.
+type SentDataMarkerHandler interface {
+	UpdateSentData(segmentId, dataCount int)
+}
+
 // batch holds pending log streams waiting to be sent to Loki, and it's used
 // to reduce the number of push requests to Loki aggregating multiple log streams
 // and entries in a single batch request. In case of multi-tenant Promtail, log
@@ -53,16 +59,6 @@ func newBatch(maxStreams int, entries ...loki.Entry) *batch {
 	return b
 }
 
-type SendDataMarkerHandler interface {
-	UpdateSentData(segmentId, dataCount int)
-}
-
-func (b *batch) reportAsSentData(h SendDataMarkerHandler) {
-	for seg, data := range b.segmentCounter {
-		h.UpdateSentData(seg, data)
-	}
-}
-
 // add an entry to the batch
 func (b *batch) add(entry loki.Entry) error {
 	b.totalBytes += len(entry.Line)
@@ -86,7 +82,8 @@ func (b *batch) add(entry loki.Entry) error {
 	return nil
 }
 
-// add an entry to the batch
+// addFromWAL adds an entry to the batch, tracking that the data being added comes from segment segmentNum read from the
+// WAL.
 func (b *batch) addFromWAL(lbs model.LabelSet, entry logproto.Entry, segmentNum int) error {
 	b.totalBytes += len(entry.Line)
 
@@ -188,10 +185,19 @@ func (b *batch) createPushRequest() (*logproto.PushRequest, int) {
 	return &req, entriesCount
 }
 
-func (b *batch) countForSegment(segment int) {
-	if curr, ok := b.segmentCounter[segment]; ok {
-		b.segmentCounter[segment] = curr + 1
+// countForSegment tracks that one data item has been read from a certain WAL segment.
+func (b *batch) countForSegment(segmentNum int) {
+	if curr, ok := b.segmentCounter[segmentNum]; ok {
+		b.segmentCounter[segmentNum] = curr + 1
 		return
 	}
-	b.segmentCounter[segment] = 1
+	b.segmentCounter[segmentNum] = 1
+}
+
+// reportAsSentData will report for all segments whose data is part of this batch, the amount of that data as sent to
+// the provided SentDataMarkerHandler
+func (b *batch) reportAsSentData(h SentDataMarkerHandler) {
+	for seg, data := range b.segmentCounter {
+		h.UpdateSentData(seg, data)
+	}
 }