--- a/component/common/loki/client/internal/marker_encoding_test.go
+++ b/component/common/loki/client/internal/marker_encoding_test.go
@@ -1,8 +1,9 @@
 package internal
 
 import (
-	"github.com/stretchr/testify/require"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
 func TestMarkerEncodingV1(t *testing.T) {