--- a/component/common/loki/client/internal/marker_file_handler.go
+++ b/component/common/loki/client/internal/marker_file_handler.go
@@ -17,14 +17,17 @@ const (
 	MarkerFileName   = "segment"
 )
 
+// MarkerFileHandler is a file-backed wal.Marker, that also allows one to write to the backing store as particular
+// segment number as the last one marked.
 type MarkerFileHandler interface {
 	wal.Marker
+
+	// MarkSegment writes in the backing file-store that a particular segment is the last one marked.
 	MarkSegment(segment int)
 }
 
 type markerFileHandler struct {
-	logger log.Logger
-
+	logger                    log.Logger
 	lastMarkedSegmentFilePath string
 }
 
@@ -32,6 +35,7 @@ var (
 	_ MarkerFileHandler = (*markerFileHandler)(nil)
 )
 
+// NewMarkerFileHandler creates a new markerFileHandler.
 func NewMarkerFileHandler(logger log.Logger, walDir string) (MarkerFileHandler, error) {
 	markerDir := filepath.Join(walDir, MarkerFolderName)
 	// attempt to create dir if doesn't exist