… component/common/loki/client/internal/marker_file_handler_test.go
… component/common/loki/client/internal/marker_handler_test.go
… component/common/loki/client/internal/metrics.go
… component/common/loki/client/manager.go
… component/common/loki/client/queue_client.go
… component/common/loki/client/queue_client_test.go
… component/common/loki/wal/watcher.go
… component/common/loki/wal/watcher_metrics.go
… component/common/loki/wal/watcher_test.go
… go.mod
… go.sum