<!--

CONTRIBUTORS GUIDE: https://github.com/grafana/agent/blob/main/docs/developer/contributing.md#updating-the-changelog

If this is your first PR or you have not contributed in a while, we recommend
taking the time to review the guide. It gives helpful instructions for
contributors around things like how to update the changelog.

-->

#### PR Description

This PR branches of https://github.com/grafana/agent/pull/5434, and applies the work done in https://github.com/ptodev/prometheus/pull/1 to improve two things:
1. WAL replay is implemented for `loki.write`, so upon an agent restart, it will replay "not sent" data saved in the WAL segments that are alive
2. Implement a "markers" mechanism that keeps track of which was the last consumed segment, consumed being read and delivered to the RW destination. This mechanisms works by creating a feedback loop between the side that reads from the WAL, and the side that sends batches.

#### TODOs
- [ ] Add tests for corrupted marker files https://github.com/grafana/agent/pull/5590#discussion_r1372979663

#### Which issue(s) this PR fixes

<!-- Uncomment the following line if you want that GitHub issue gets automatically closed after merging the PR -->
<!-- Fixes #issue_id -->

#### Notes to the Reviewer

#### PR Checklist

<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [ ] CHANGELOG.md updated
- [ ] Documentation added
- [ ] Tests updated
- [ ] Config converters updated