@thepalbi 2023-10-26
@ptodev went through ~half the comments, and rebased over the latest changes in https://github.com/grafana/agent/pull/5434, which is already approved.