@thepalbi 2023-11-01
@ptodev @mattdurham just added on top of the previous queue client benchmarks, once comparing the nil marker and real marker implementations. No significant diff whatsoever:
```
goos: darwin
goarch: arm64
pkg: github.com/grafana/agent/component/common/loki/client
                                                                    │ queue_nil_marker_handler │       queue_marker_handler       │
                                                                    │          sec/op          │   sec/op    vs base              │
ClientImplementations/100_entries,_single_series,_no_batching-10                    1.001 ± 0%   1.001 ± 0%       ~ (p=0.394 n=6)
ClientImplementations/100k_entries,_100_series,_default_batching-10                 4.292 ± 1%   4.336 ± 1%  +1.02% (p=0.026 n=6)
geomean                                                                             2.073        2.084       +0.50%

                                                                    │ queue_nil_marker_handler │        queue_marker_handler        │
                                                                    │           B/op           │     B/op      vs base              │
ClientImplementations/100_entries,_single_series,_no_batching-10                  861.0Ki ± 5%   861.6Ki ± 5%       ~ (p=0.615 n=6)
ClientImplementations/100k_entries,_100_series,_default_batching-10               788.9Mi ± 0%   788.0Mi ± 0%  -0.12% (p=0.002 n=6)
geomean                                                                           25.75Mi        25.75Mi       -0.02%

                                                                    │ queue_nil_marker_handler │       queue_marker_handler        │
                                                                    │        allocs/op         │  allocs/op   vs base              │
ClientImplementations/100_entries,_single_series,_no_batching-10                   11.26k ± 1%   11.28k ± 0%       ~ (p=0.058 n=6)
ClientImplementations/100k_entries,_100_series,_default_batching-10                10.80M ± 0%   10.80M ± 0%  -0.01% (p=0.002 n=6)
geomean                                                                            348.8k        349.1k       +0.09%
```